[I25-261] feat : 모달 생성 로직을 훅에서 유틸 함수로 리팩토링
