+++ src/app/components/auth/Account.tsx
@@ -1,13 +1,14 @@
 'use client';
 
 import { createClient } from '@/services/supabase/client';
-import { useEffect } from 'react';
+import { useEffect, useCallback } from 'react';
 import {Dropdown, DropdownItem} from '../dropdown/Dropdown';
-import { useCreateProject } from '@/utils/hook/useCreateProject';
-import { useCreateProfile } from '@/utils/hook/useCreateProfile';
+import { useModal } from '@/app/components/modal';
 import { useCurrentUser } from '@/utils/hook/useCurrentUser';
 import { useProfileLink } from '@/utils/hook/useProfileLink';
-import { useGoogleLogin } from '@/utils/hook/useGoogleLogin';
+import { openProjectModal } from '@/utils/modal/openProjectModal';
+import { openProfileModal } from '@/utils/modal/openProfileModal';
+import { openGoogleLogin } from '@/utils/auth/googleLogin';
 import { checkProfileExistence } from '@/services/profile/getProfileApi.client';
 import Image from 'next/image';
 import Link from 'next/link';
@@ -16,15 +17,29 @@ const Account = () => {
   const supabase = createClient();
   const currentUser = useCurrentUser();
   const profileLink = useProfileLink();
-  const handleMakeProfile = useCreateProfile();
-  const handleMakeProject = useCreateProject();
-  const handleGoogleLogin = useGoogleLogin();
+  const { openModal, closeModal } = useModal();
+
+  const handleMakeProfile = useCallback(async () => {
+    if (!currentUser?.id) {
+      alert('로그인이 필요합니다.');
+      return;
+    }
+    await openProfileModal(currentUser.id, openModal, closeModal);
+  }, [currentUser?.id, openModal, closeModal]);
+
+  const handleMakeProject = useCallback(async () => {
+    if (!currentUser?.id) {
+      alert('로그인이 필요합니다.');
+      return;
+    }
+    await openProjectModal(currentUser.id, openModal, closeModal);
+  }, [currentUser?.id, openModal, closeModal]);
 
   useEffect(() => {
     if (!currentUser?.id) return;
     checkProfileExistence(currentUser.id).then((exists) => {
       if (!exists) {
-        handleMakeProfile();
+        void handleMakeProfile();
       }
     });
   }, [currentUser?.id, handleMakeProfile]);
@@ -56,7 +71,7 @@ const Account = () => {
     </Dropdown>
   ) : (
     <button
-      onClick={handleGoogleLogin}
+      onClick={openGoogleLogin}
       className="px-2 py-1 bg-black text-caption text-white rounded-full transition-colors"
     >
       로그인
